requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.6
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.29
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
proto-plus==1.24.0
protobuf==4.21.12
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
